Solution to Calculate and Print the Average of Three Numbers

+++ Calculate and Print the Average of Three Numbers@@6788d5012db25efec946f944/code.c
@@ -1,8 +1,8 @@
 #include <stdio.h>
 int main() {
-    int a,b,c,d;
-    scanf("%d %d %d", &a,&b,&c);
-    d=a+b+c/3.0;
+    float a,b,c,d;
+    scanf("%f %f %f", &a,&b,&c);
+    d=a+b+c/3;
     printf("Average: %.2f",d);
     return 0;
 }
